Add a way to piecemeal construct an fc::url
fc::mutable_url was never implemented; so remove the deceiving declaration of it. Add a way to piecemeal construct an fc::url by a long argument list in its place. I need to construct an fc::url with something that isn't quite a real URL so I need a way to set each piece manually.